Organised C++ directory
